move components and layouts into pages
