+++ web/src/pages/components/EmailButton.vue
