+++ web/src/pages/components/ResultView.vue
