+++ web/src/pages/components/TranslatorCheck.vue
