+++ web/src/pages/layouts/ReaderLayout.vue
