+++ web/src/pages/layouts/components/RobotIcon.vue
