+++ web/src/pages/workspace/components/JobWorker.vue
@@ -8,7 +8,7 @@ import {
 import { useMessage } from 'naive-ui';
 import { computed, ref } from 'vue';
 
-import TranslateTask from '@/components/TranslateTask.vue';
+import TranslateTask from '@/pages/components/TranslateTask.vue';
 import {
   GptWorker,
   SakuraWorker,
